Make `Fixed` obey `Enum` laws
`Enum` states that `succ` means add 1, and `pred` means subtract 1.
`enumFrom` and `enumFromTo` behave similarly. In fact, there are more
subtleties regarding both fractional numeric types as well as bounded
types. See the Haskell Report 2010 section 6.3.4 "The Enum Class", as
well as the documentation of the `Enum` class in the Haskell Prelude for
details.